Add controllers for the products
