+++ routes/admin.js
@@ -1,27 +1,12 @@
 const express = require('express');
-const path = require('path');
+// const path = require('path');
+// const rootDir = require('../util/path');
 
-const rootDir = require('../util/path');
-
+const productsController = require('../controllers/products');
 const router = express.Router();
 
-const products = [];
+router.get('/add-product', productsController.getAddProduct);
 
-router.get('/add-product', (req, res, next) => {
-  res.render('ejs/add-product', {
-    products,
-    pageTitle: 'Add Product',
-    path: '/admin/add-product',
-    activeAddProduct: true,
-    productCss: true,
-    formsCss: true,
-  });
-});
+router.post('/add-product', productsController.postAddProduct);
 
-router.post('/add-product', (req, res, next) => {
-  products.push({ title: req.body.title });
-
-  res.redirect('/');
-});
-
-module.exports = { routes: router, products };
+module.exports = router;
