+++ routes/shop.js
@@ -1,23 +1,8 @@
 const express = require('express');
-const path = require('path');
-
-const rootDir = require('../util/path');
-
-const adminData = require('./admin');
+const productsController = require('../controllers/products');
 
 const router = express.Router();
 
-router.get('/', (req, res, next) => {
-  const products = adminData.products;
-
-  res.render('ejs/shop', {
-    products,
-    pageTitle: 'Shop',
-    path: '/',
-    hasProducts: products.length > 0,
-    activeShop: true,
-    productCss: true,
-  });
-});
+router.get('/', productsController.getProducts);
 
 module.exports = router;
